requirements.txt
SpeechRecognition
pyttsx3
requests
wolframalpha
pywhatkit
pyautogui
pyjokes
Pillow
selenium
webdriver-manager
psutil
GPUtil
screen-brightness-control
pycaw
comtypes
PyQt5
pywin32
lxml
beautifulsoup4
wikipedia
geopy
geocoder
pipwin
elevenlabs
PyPDF2
python-docx
pytz
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
